+++ lib/src/legacy/value/index.ts
@@ -1,0 +1,9 @@
+// Copyright 2022 Google LLC. Use of this source code is governed by an
+// MIT-style license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import {SassBooleanInternal} from '../../value/boolean';
+import {SassNull} from '../../value/null';
+
+export const Boolean = SassBooleanInternal;
+export const Null = SassNull;