+++ lib/src/value/boolean.ts
@@ -19,9 +19,22 @@ export interface SassBoolean extends Value {
 const trueHash = hash(true);
 const falseHash = hash(false);
 
-class SassBooleanInternal extends Value implements SassBoolean {
+export class SassBooleanInternal extends Value implements SassBoolean {
+  // Whether callers are allowed to construct this class. This is set to
+  // `false` once the two constants are constructed so that the constructor
+  // throws an error for future calls, in accordance with the legacy API.
+  static constructionAllowed = true;
+
   constructor(private readonly valueInternal: boolean) {
     super();
+
+    if (!SassBooleanInternal.constructionAllowed) {
+      throw (
+        "new sass.types.Boolean() isn't allowed.\n" +
+        'Use sass.types.Boolean.TRUE or sass.types.Boolean.FALSE instead.'
+      );
+    }
+
     Object.freeze(this);
   }
 
@@ -48,6 +61,15 @@ class SassBooleanInternal extends Value implements SassBoolean {
   toString(): string {
     return this.value ? 'sassTrue' : 'sassFalse';
   }
+
+  // Legacy API support
+
+  static TRUE: SassBooleanInternal;
+  static FALSE: SassBooleanInternal;
+
+  getValue(): boolean {
+    return this.value;
+  }
 }
 
 /** The singleton instance of SassScript true. */
@@ -55,3 +77,9 @@ export const sassTrue = new SassBooleanInternal(true);
 
 /** The singleton instance of SassScript false. */
 export const sassFalse = new SassBooleanInternal(false);
+
+// Legacy API support
+SassBooleanInternal.constructionAllowed = false;
+
+SassBooleanInternal.TRUE = sassTrue;
+SassBooleanInternal.FALSE = sassFalse;